pyproject.toml
[tool.poetry]
name = "phoenix-site"
version = "0.1.0"
description = "phoenix-studio pres."
authors = ["macalistervadim <macalistervadim@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
python-dotenv = "^1.0.1"
django-debug-toolbar = "^4.4.2"
psycopg = "^3.1.19"
psycopg-binary = "^3.1.19"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
flake8-absolute-import = "^1.0.0.2"
flake8-commas = "^4.0.0"
flake8-import-order = "^0.18.2"
flake8-print = "^5.0.0"
flake8-quotes = "^3.4.0"
flake8-use-pathlib = "^0.3.0"
pep8-naming = "^0.14.1"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
